User 클래스에서 UserThread 분리작업
